requirements.txt
numpy==1.23.0
Pillow
matplotlib
chromadb==0.5.5
open-clip-torch
tqdm
selenium
#chromium-browser
#chromium-chromedriver